:art: Refactor code
