+++ src/behaviors/deterministic/ascendent.py
@@ -1,0 +1,23 @@
+from typing import List
+
+from src.behaviors.behaviors import Behavior
+from src.player import Player
+
+
+class DetAscPlayer(Player):
+
+    def __init__(self, id_player: int):
+        super().__init__(id_player)
+        self.behavior = Behavior.DETASC
+
+    def play(self, players: List[Player]):
+        self.kill_asc(players)
+
+    def kill_asc(self, targets: List['Player']):
+        my_index = targets.index(self)
+        target = targets[my_index + 1] if my_index != len(targets) - 1 else targets[0]
+        while target is self:  # cas ou targets[0] is me
+            i = 1
+            target = targets[i + 1]
+        target.killed = True
+        self.kills += 1
