+++ src/player.py
@@ -21,3 +21,11 @@ class Player(ABC):
 
     def __repr__(self):
         return f" {self.id} : {self.kills}"
+
+class IdlePlayer(Player):
+    def __init__(self, id_player: int):
+        super().__init__(id_player)
+        self.behavior = Behavior.IDLE
+
+    def play(self, players: List[Player]):  # do nothing
+        pass
